+++ back/app.js
@@ -5,29 +5,8 @@ const app = express();
 // Mise en place de CORS 
 const cors = require('cors');
 
-// Mise en place de body-parser
-//const bodyParser = require('body-parser');
-
 // Mise en place du path pour gérer l'url d'accès à l'image avec le dirname pour MAJ auto
 const path = require('path');
-
-// Mise en place de multer pour l'upload des images
-const multer = require('multer');
-const storage = multer.diskStorage({
-    destination: "images/", 
-    filename: function (req, file, cb){
-    cb(null, creationFilename(req, file))
-    }
-});
-
-// Fonction de création du nom d'image unique
-function creationFilename(req, file){
-    const fileName = `${Date.now()}-${file.originalname}`.replace(/\s/g, "-")
-    file.fileName = fileName
-    return fileName
-};
-
-const upload = multer({ storage: storage});
 
 /* - - - - - BASSE DE DONNEES - - - - - */
 // Import du fichier JS mongo
@@ -38,7 +17,7 @@ require('./mongo');
 const {creationUser, loginUser} = require('./controllers/user');
 
 // Récupération de la fonction de recup et, de création des sauces
-const {recupSauces, creationSauces} = require('./controllers/sauces');
+const {recupSauces, creationSauces, recupSauceDepuisId} = require('./controllers/sauces');
 
 /* - - - - - MIDDLEWARES - - - - - */
 // Sert à se connecter avec Express, pour faire des requetes au serveur et, recevoir des réponses
@@ -49,6 +28,9 @@ app.use(express.json());
 
 // Récupération du middleware d'authentification de l'utilisateur
 const {authentUser} = require('./middlewares/auth');
+
+// Mise en place du multer.js pour récupérer le upload
+const {upload} = require('./middlewares/multer');
 
 /* - - - - - ROUTES - - - - - */
 // Post pour la création d'un utilisateur au Signup
@@ -63,9 +45,11 @@ app.get('/api/sauces', authentUser, recupSauces);
 // Post pour l'ajout, la création de sauces et l'upload de l'image (Authentification User requis)
 app.post('/api/sauces', authentUser, upload.single("image"), creationSauces);
 
-// Test pour vérifier le bon accés au port 3000
-app.get("/", (req, res) => res.send("Test d'affichage"));
+// Get lors du clic sur une sauce
+app.get ('/api/sauces/:id', authentUser, recupSauceDepuisId)
 
+// Requete pour vérifier le bon accés au port 3000
+app.get("/", (req, res) => res.send("Test d'affichage"));
 
 // MIDDLEWARE - Sert à rendre accesible l'image uploadée depuis l'url
 app.use("/images", express.static(path.join(__dirname, "images")));
